made email function into module to export in app.js
