+++ baseSite/README.md
@@ -3,3 +3,5 @@ MongoTest - add, delete, select records from DB
 npm start - in root to start the bin/www
 
 1st - mongo needs to be installed. npm install needs ran to install package.json
+
+.env file needs to be created in 'envMonds/.env'
