--- baseSite/calendarTest/layout.html
@@ -1,144 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<style>
-* {box-sizing: border-box;}
-ul {list-style-type: none;}
-body {font-family: Verdana, sans-serif;}
-
-.month {
-  padding: 70px 25px;
-  width: 100%;
-  background: #1abc9c;
-  text-align: center;
-}
-
-.month ul {
-  margin: 0;
-  padding: 0;
-}
-
-.month ul li {
-  color: white;
-  font-size: 20px;
-  text-transform: uppercase;
-  letter-spacing: 3px;
-}
-
-.month .prev {
-  float: left;
-  padding-top: 10px;
-}
-
-.month .next {
-  float: right;
-  padding-top: 10px;
-}
-
-.weekdays {
-  margin: 0;
-  padding: 10px 0;
-  background-color: #ddd;
-}
-
-.weekdays li {
-  display: inline-block;
-  width: 13.6%;
-  color: #666;
-  text-align: center;
-}
-
-.days {
-  padding: 10px 0;
-  background: #eee;
-  margin: 0;
-}
-
-.days li {
-  list-style-type: none;
-  display: inline-block;
-  width: 13.6%;
-  text-align: center;
-  margin-bottom: 5px;
-  font-size:12px;
-  color: #777;
-}
-
-.days li .active {
-  padding: 5px;
-  background: #1abc9c;
-  color: white !important
-}
-
-/* Add media queries for smaller screens */
-@media screen and (max-width:720px) {
-  .weekdays li, .days li {width: 13.1%;}
-}
-
-@media screen and (max-width: 420px) {
-  .weekdays li, .days li {width: 12.5%;}
-  .days li .active {padding: 2px;}
-}
-
-@media screen and (max-width: 290px) {
-  .weekdays li, .days li {width: 12.2%;}
-}
-</style>
-</head>
-<body>
-
-<h1>CSS Calendar</h1>
-
-<div class="month">      
-  <ul>
-    <li class="prev">&#10094;</li>
-    <li class="next">&#10095;</li>
-    <li>
-      <p id="setMonth"></p><br>
-      <span id="setYear"></span>
-    </li>
-  </ul>
-</div>
-
-<script>
-    const monthNames = ["January", "February", "March", "April", "May", "June",
-                        "July", "August", "September", "October", "November", "December"];
-    const date1 = new Date();
-    console.log(monthNames[date1.getMonth()]);
-    month1 = (monthNames[date1.getMonth()]);
-    year1 = date1.getFullYear();
-    document.getElementById("setMonth").innerHTML = (monthNames[date1.getMonth()]);
-    document.getElementById("setYear").innerHTML = date1.getFullYear();
-</script>
-
-<script>
-    document.write("<ul class=\"weekdays\">");
-        document.write("<li>Mon</li>");
-        document.write("<li>Tue</li>");
-        document.write("<li>Wed</li>");
-        document.write("<li>Thur</li>");
-        document.write("<li>Fri</li>");
-        document.write("<li>Sat</li>");
-        document.write("<li>Sun</li>");
-    document.write("</ul>");
-    </script>
-
-<script>
-    date2 = new Date();
-    date3 = new Date()
-    document.write(date3);
-    days = new Date(year1, date2.getMonth() + 1, 0);
-    document.write(days);
-    test1 = String(days);
-    test1 = test1.substring(8,10)
-    document.write(test1);
-
-    document.write("<ul class=\"days\">")  
-    for(i = 1;i <= test1; i++ ) {
-        document.write("<li>"+i+"</li>");
-    }
-    document.write("</ul>");
-</script>
-
-</body>
-</html>
