+++ baseSite/envMods/envConfig.js
@@ -1,4 +1,4 @@
-const dotenv = require('dotenv').config({path:__dirname + '/.env'});
+const dotenv = require('dotenv').config({path: __dirname + '/.env'});
 
 // dotenv.config();
